(수정) 탭스크린, 디테일페이지, sortList value (추가) 캘린더
feature-issue-#41
