+++ pages/navigation/TabNavigator.tsx
@@ -2,37 +2,38 @@ import * as React from 'react'
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs'
 import MainListNavigation from './MainListNavigation'
 import MyPageNavigation from './MypageNavigation'
-import Detail from '../screens/Detail'
 import Ionicons from "react-native-vector-icons/Ionicons"
+import { Image } from "react-native"
+
+
 
 export default function TabNavigator () {
     const Tab = createBottomTabNavigator()
 
     return (
-        <Tab.Navigator initialRouteName="MainList" screenOptions={{tabBarActiveTintColor:'red', tabBarActiveBackgroundColor:'beige'}} >
+        <Tab.Navigator initialRouteName="MainList" screenOptions={{tabBarActiveTintColor:'#518099'}} >
             <Tab.Screen name="MainList" 
-                component={MainListNavigation} 
-                options={{headerShown:false, tabBarIcon:()=>(<Ionicons 
-                        name="list-outline"
-                        size={20}/>
-                )}} 
+                component={MainListNavigation}
+                options={{headerShown:false, tabBarIcon:({size})=> {
+                    return (
+                        <Image
+                            style={{width:size, height:size }}
+                            source={require('../../public/images/List/checkList.png')}
+                        />
+                    )
+                }}}
             />
-            <Tab.Screen name="MyPage" 
+            <Tab.Screen name="MyPage"
                 component={MyPageNavigation} 
-                options={{tabBarIcon:()=>(<Ionicons
-                name="apps-outline"
-                size={20}/>
-
-                )}}
-            />
-            <Tab.Screen name="Detail" 
-                component={Detail} 
-                options={{headerShown:false, tabBarIcon:()=>(<Ionicons 
-                        name="list-outline"
-                        size={20}/>
-                )}} 
+                options={{tabBarIcon:({size})=> {
+                    return (
+                        <Image
+                            style={{width:size, height:size }}
+                            source={require('../../public/images/List/user.png')}
+                        />
+                    )
+                }}}
             />
         </Tab.Navigator>
-        
     )
 }
