+++ pages/screens/Calendar/index.tsx
@@ -1,10 +1,8 @@
-import React, { useContext } from 'react'
+import React from 'react'
 import CalendarView from "../../../src/component/Calendar/CalendarView"
-import LogContext from "../../../contexts/LogContext"
 
 export default function CalendarScreen({navigation}){
 
-
-    return <CalendarView/>
-}
+    return <CalendarView />
+};
