+++ src/component/Picker/ImportantPicker.tsx
@@ -32,10 +32,10 @@ export default function ImportantPicker(props){
         mode='dialog'
         style={{borderWidth:1, borderStyle:'solid'}}
         >
-        <Picker.Item label='    🔴  (1순위)   ' value='1' />
-        <Picker.Item label='    🟠  (2순위)    ' value='2' />
-        <Picker.Item label='    🟢  (3순위)    ' value='3' />
-        <Picker.Item label='    🔵  (4순위)    ' value='4' />
+        <Picker.Item label='🔴 (1순위)' value='🔴 (1순위)' />
+        <Picker.Item label='🟠 (2순위)' value='🟠 (2순위)' />
+        <Picker.Item label='🟢 (3순위)' value='🟢 (3순위)' />
+        <Picker.Item label='🔵 (4순위)' value='🔵 (4순위)' />
       </Picker>
     </Container>
     </Wrapper>
